% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionOR.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Odds-Ratio and Confidence Interval for Logistic Regression predictors}
\usage{
OR_95CI(formula, data, siglevel, roundto)
}
\arguments{
\item{X}{description predictor variable}
}
\value{
odds ratio and 95% confidence interval
}
\description{
input a function of binary variable Y given Xn predictors and
return output of odds ratio and confidence interval for each predictor variable
}
\examples{
OR_95CI()
}
\author{
Mohsyn Imran Malik
}
