% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 100 rows and 3 variables: {
y = a dependent binary outcome variable
xi where i = {1,2} are independent explanatory variables of y
}
}
\usage{
toydata
}
\description{
Sample data with binary outcome and explanatory covariates
}
\keyword{datasets}
